requirements.txt
#
# Core runtime
#
annotated-types==0.7.0
packaging==26.0
platformdirs==4.9.4
setuptools==82.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
wheel==0.46.3

#
# Web / app framework
#
asgiref==3.11.1
Django==6.0.2
sqlparse==0.5.5
tzdata==2025.3

#
# Config / validation / reporting
#
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
rich==14.3.3

#
# Art / image / procedural generation
#
noise==1.2.2
numpy==2.4.2
opencv-python==4.13.0.92
pillow==12.1.1
pygame-ce==2.5.6
pytweening==1.2.0

#
# Audio / signal analysis
#
audioop-lts==0.2.2
audioread==3.1.0
decorator==5.2.1
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
msgpack==1.1.2
numba==0.64.0
pooch==1.9.0
scikit-learn==1.8.0
scipy==1.17.1
soundfile==0.13.1
soxr==1.0.0
threadpoolctl==3.6.0

#
# Asset / file I/O / graph utilities
#
cffi==2.0.0
networkx==3.6.1
pycparser==3.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0

#
# HTTP / external data helpers
#
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
requests==2.32.5
urllib3==2.6.3
